WIP: matcha: Updated files

+++ wip/matcha/backend/src/main/java/ft/framework/validation/constraint/validator/PositiveOrZeroValidator.java
@@ -11,7 +11,7 @@ public class PositiveOrZeroValidator implements ConstraintValidator<PositiveOrZe
 			return true;
 		}
 		
-		return value.intValue() >= 0;
+		return value.longValue() >= 0;
 	}
 	
 }
